refactor : Factory 정리
