+++ screennameviewer/compose/src/main/java/com/donglab/screennameviewer/config/ScreenNameViewerConstants.kt
@@ -1,0 +1,6 @@
+package com.donglab.screennameviewer.config
+
+internal object ScreenNameViewerConstants {
+
+    const val FRAGMENT_LAYOUT_TAG = "ScreenNameViewer_FragmentLayout"
+}
